+++ Maracas.Lib/Models/Order.cs
@@ -6,15 +6,10 @@ namespace Maracas.Lib.Models
 {
     public class Order
     {
-        public long Id { get; set; }
-
-        public long ClientId { get; set; }
-
+        public int Id { get; set; }
+        public int ClientId { get; set; }
         public Client Client { get; set; }
-
         public List <Product> Products { get; set; }
-
-
 
     }
 }
